refactor: migrate calory and receipt assistant logic to usecases and remove deprecated code
