+++ openclaw-platform/src/usecases/receipt-assistant/select-receipt-media.ts
